+++ docs/release_notes.rst
@@ -4,8 +4,8 @@ Release Notes
 **pydocstyle** version numbers follow the
 `Semantic Versioning <http://semver.org/>`_ specification.
 
-x.x.x - Current Development Version
------------------------------------
+Current Development Version
+---------------------------
 
 * The error code D300 is now also being reported if a docstring has
   uppercase literals (``R`` or ``U``) as prefix (#176).